requirements-compat.txt
Beaker>=1.12.0
beautifulsoup4>=4.12.0
Django>=3.2.0
Mako>=1.2.0
Markdown>=3.4.0
MarkupSafe>=2.1.0
PyYAML>=6.0
textblob>=0.17.0
simplejson>=3.17.0
bottle>=0.12.0
cohere>=4.37
defusedxml>=0.7.0
django-cache-machine>=1.0.0
django-fancy-cache>=0.10.0
django-haystack>=3.2.0
django-tastypie>=0.14.0
iowait>=0.2
lxml>=4.9.0
mimeparse>=0.1.3
mistralai>=1.5.1
mock>=4.0.0
psutil>=5.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-ptrace>=0.9.0
requests>=2.28.0
shortuuid>=1.0.0
six>=1.16.0
django-cors-headers>=4.0.0